Refactor settings panel for reuse in CPT
