+++ packages/block-library/form/index.js
@@ -20,6 +20,8 @@ import edit from './edit';
 import { POST_TYPE } from '../shared/constants';
 import { form } from '../shared/icons';
 
+import './document';
+
 import './index.scss';
 import './style.scss';
